work-it/requirements.txt
discord==1.0.1
discord.py==1.4.1
python-dotenv==0.14.0
